third_party/WebKit/LayoutTests/inspector/sources/debugger-pause/debugger-pause-on-exception-expected.txt

CONSOLE ERROR: line 9: Uncaught ReferenceError: unknown_var is not defined
Tests that pause on exception works.

Script source was shown.
Script execution paused.
Call stack:
    0) throwAnException (debugger-pause-on-exception.html:9)
    1) handleClick (debugger-pause-on-exception.html:14)
Paused on exceptionReferenceError: unknown_var is not defined
Script execution resumed.
